refactor: move Goto-related stuff into a new plugin
Identified during this refactor is an issue where the "seconds since
last reveal" timer needs to be reset in multiple places, ie anywhere
that transitions back to the playback state. The importance is to reset
it directly before transitioning to the playback state and no earlier.
The defect seen otherwise is a burst of text being revealed all at once
when starting a new passage.

A possible solution is noted in comments in the diff, but to summarize
we could isolate parts of the BillboardData component which signal the
reset is required in their own component. That way, the playback system
can watch this one specific component for changes and reset the timer at
the head of the system.
